feat(interface): obliterating redux and removing unused / legacy code
