+++ apps/namada-interface/package.json
@@ -7,7 +7,6 @@
   "license": "MIT",
   "private": true,
   "dependencies": {
-    "@reduxjs/toolkit": "^1.8.0",
     "@types/invariant": "^2.2.37",
     "@types/react-paginate": "^7.1.2",
     "bignumber.js": "^9.1.1",
@@ -20,7 +19,6 @@
     "framer-motion": "^11.0.28",
     "invariant": "^2.2.4",
     "jotai": "^2.6.3",
-    "jotai-redux": "^0.2.1",
     "lodash.debounce": "^4.0.8",
     "next-qrcode": "^2.0.0",
     "path": "^0.12.7",
@@ -30,12 +28,8 @@
     "react-modal": "^3.15.1",
     "react-paginate": "^8.2.0",
     "react-qr-reader": "2.1.2",
-    "react-redux": "^7.2.6",
     "react-router-dom": "^6.0.0",
     "react-scripts": "5.0.1",
-    "redux-persist": "^6.0.0",
-    "redux-persist-transform-encrypt": "^3.0.1",
-    "redux-thunk": "^2.4.1",
     "stream": "^0.0.2",
     "styled-components": "^5.3.3",
     "tailwind-merge": "^2.3.0",
@@ -96,7 +90,6 @@
     "@types/react-dom": "^17.0.11",
     "@types/react-modal": "^3.13.1",
     "@types/react-qr-reader": "^2.1.4",
-    "@types/redux-mock-store": "^1.0.3",
     "@types/styled-components": "^5.1.22",
     "@vitejs/plugin-react": "^4.2.1",
     "autoprefixer": "^10.4.16",
@@ -116,7 +109,6 @@
     "local-cors-proxy": "^1.1.0",
     "postcss": "^8.4.32",
     "postcss-loader": "^7.3.3",
-    "redux-mock-store": "^1.5.4",
     "style-loader": "^3.3.1",
     "tailwindcss": "^3.4.0",
     "ts-jest": "^29.0.5",
