+++ apps/namada-interface/src/App/Bridge/Bridge.tsx
@@ -1,6 +1,5 @@
 import { Heading, NavigationContainer } from "@namada/components";
 import IBCTransfer from "App/Token/IBCTransfer/IBCTransfer";
-import { BridgeContainer } from "./Bridge.components";
 // import { useState } from "react";
 // import { EthereumBridge } from "App/Token";
 
@@ -10,7 +9,7 @@ export const Bridge = (): JSX.Element => {
   // const [activeTab, setActiveTab] = useState(tabs[0]);
 
   return (
-    <BridgeContainer>
+    <>
       <NavigationContainer>
         {/* <Heading level="h1">Bridge</Heading> */}
         <Heading level="h1">IBC Transfer</Heading>
@@ -29,6 +28,6 @@ export const Bridge = (): JSX.Element => {
       {/* </TabsGroup> */}
       {/* {activeTab == tabs[0] ? <IBCTransfer /> : <EthereumBridge />} */}
       <IBCTransfer />
-    </BridgeContainer>
+    </>
   );
 };
