+++ apps/namada-interface/src/App/Common/AnimatedTransition.tsx
@@ -1,0 +1,20 @@
+import { motion } from "framer-motion";
+
+export const AnimatedTransition = (props: {
+  children: React.ReactNode;
+  elementKey: string;
+}): JSX.Element => {
+  const { children, elementKey } = props;
+  return (
+    <motion.div
+      className="flex justify-center h-full w-full"
+      key={elementKey}
+      initial={{ opacity: 0 }}
+      animate={{ opacity: 1 }}
+      exit={{ opacity: 0 }}
+      transition={{ duration: 0.5 }}
+    >
+      {children}
+    </motion.div>
+  );
+};
