--- apps/namada-interface/src/config/index.ts
@@ -1,7 +1,0 @@
-import ApiConfig from "./api";
-
-const Config = {
-  api: ApiConfig,
-};
-
-export default Config;
