+++ apps/namada-interface/src/index.tsx
@@ -2,12 +2,10 @@ import { init as initShared } from "@namada/shared/src/init-inline";
 import { SdkProvider } from "hooks/useSdk";
 import React from "react";
 import { createRoot } from "react-dom/client";
-import { Provider } from "react-redux";
 import { RouterProvider } from "react-router-dom";
 import { getRouter } from "./App/AppRoutes";
 import reportWebVitals from "./reportWebVitals";
 import { ExtensionEventsProvider, IntegrationsProvider } from "./services";
-import { store } from "./store/store";
 
 import "@namada/components/src/base.css";
 import "./tailwind.css";
@@ -21,11 +19,9 @@ if (container) {
       <React.StrictMode>
         <IntegrationsProvider>
           <SdkProvider>
-            <Provider store={store}>
-              <ExtensionEventsProvider>
-                <RouterProvider router={getRouter()} />
-              </ExtensionEventsProvider>
-            </Provider>
+            <ExtensionEventsProvider>
+              <RouterProvider router={getRouter()} />
+            </ExtensionEventsProvider>
           </SdkProvider>
         </IntegrationsProvider>
       </React.StrictMode>
