--- apps/namada-interface/src/slices/index.ts
@@ -1,2 +1,0 @@
-export { default as channelsReducer } from "./channels";
-export { default as settingsReducer } from "./settings";
