--- apps/namada-interface/src/store/index.ts
@@ -1,3 +1,0 @@
-export { useAppDispatch, useAppSelector } from "./hooks";
-export { persistor, reduxStoreAtom, store } from "./store";
-export type { RootState } from "./store";
